Refactor: Extract composables and utility functions
- I extracted NFC logic from App.vue to useNfc.ts
- I extracted README modal logic from App.vue to useReadme.ts
- I refactored App.vue to use the new composables.
- I extracted getDefaultProfileData from IdCard.vue to profile.ts utility.
- I extracted IdCard system info and barcode logic to composables.
- I extracted SinForm license management to a composable.
- I ensured the application builds successfully after changes.
